pyproject.toml
[tool.poetry]
name = "microwave"
version = "0.1.0"
description = ""
authors = ["Piotr Michalik <piotrmichalik.it@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.100.0"
gunicorn = "^21.2.0"
pydantic-settings = "^2.0.2"
pydantic = "^2.0.3"
redis = "^4.6.0"
jinja2 = "^3.1.2"
uvicorn = { extras = ["standard"], version = "^0.23.1" }
pyjwt = "^2.8.0"
black = "^23.7.0"
pytest = "^7.4.0"
httpx = "^0.24.1"
pytest-env = "^0.8.2"


[tool.poetry.group.dev.dependencies]
setuptools = "^68.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
